refactor : pawn 관련 로직 메소드 분리
