+++ src/test/java/chess/domain/piece/pawn/NotMovedPawnStrategyTest.java
@@ -15,13 +15,13 @@ import chess.domain.Team;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 
-class MovedPawnTest {
+class NotMovedPawnStrategyTest {
 	private Piece whitePawn;
 	private Map<Position, Team> boardDto;
 
 	@BeforeEach
 	void setUp() {
-		whitePawn = MovedPawn.of(Team.WHITE, Position.of("c4"));
+		whitePawn = NotMovedPawn.of(Team.WHITE, Position.of("c4"));
 		boardDto = new HashMap<>();
 		boardDto.put(Position.of("c4"), Team.WHITE);
 	}
@@ -63,12 +63,10 @@ class MovedPawnTest {
 	}
 
 	@Test
-	@DisplayName("폰의 기본전략에서는, 앞으로 두칸 전진할 수 없습니다.")
+	@DisplayName("움직이지 않은 폰은 두칸을 이동할 수 있습니다.")
 	void canNotMoveFrontDouble() {
-		assertThatThrownBy(() -> {
-			whitePawn.move(Position.of("c4"), Position.of("c6"), boardDto);
-		}).isInstanceOf(IllegalArgumentException.class)
-			.hasMessageMatching("도달할 수 없는 거리입니다.");
+		assertThat(whitePawn.move(Position.of("c4"), Position.of("c6"), boardDto))
+			.isInstanceOf(MovedPawn.class);
 	}
 
 	@Test
